api/requirements.txt
fastapi
gunicorn
uvicorn
python-dotenv
Jinja2
pytest
requests
httpx
pre-commit
sqlalchemy
psycopg2-binary
sqlalchemy_utils
argon2_cffi
fastapi_jwt
celery
redis
alembic
flower